+++ tests/text_area/test_edit_via_bindings.py
@@ -361,6 +361,18 @@ async def test_delete_word_left(selection, expected_result, final_selection):
         assert text_area.text == expected_result
         assert text_area.selection == final_selection
 
+    # Repeat with ctrl+backspace binding
+    app = TextAreaApp()
+    async with app.run_test() as pilot:
+        text_area = app.query_one(TextArea)
+        text_area.load_text("  012 345 6789")
+        text_area.selection = selection
+
+        await pilot.press("ctrl+backspace")
+
+        assert text_area.text == expected_result
+        assert text_area.selection == final_selection
+
 
 @pytest.mark.parametrize(
     "selection,expected_result,final_selection",
@@ -424,11 +436,11 @@ async def test_delete_word_left_at_line_start():
 @pytest.mark.parametrize(
     "selection,expected_result,final_selection",
     [
-        (Selection.cursor((0, 0)), "012 345 6789", Selection.cursor((0, 0))),
+        (Selection.cursor((0, 0)), " 345 6789", Selection.cursor((0, 0))),
         (Selection.cursor((0, 4)), "  01 345 6789", Selection.cursor((0, 4))),
-        (Selection.cursor((0, 5)), "  012345 6789", Selection.cursor((0, 5))),
+        (Selection.cursor((0, 5)), "  012 6789", Selection.cursor((0, 5))),
         (Selection.cursor((0, 14)), "  012 345 6789", Selection.cursor((0, 14))),
-        # When non-empty selection, "delete word right" just deletes the selection
+        # # When non-empty selection, "delete word right" just deletes the selection
         (Selection((0, 4), (0, 11)), "  01789", Selection.cursor((0, 4))),
     ],
 )
@@ -439,9 +451,10 @@ async def test_delete_word_right(selection, expected_result, final_selection):
         text_area.load_text("  012 345 6789")
         text_area.selection = selection
 
-        await pilot.press("ctrl+f")
+        await pilot.press("alt+delete")
 
         assert text_area.text == expected_result
+        print(final_selection)
         assert text_area.selection == final_selection
 
 
@@ -452,7 +465,7 @@ async def test_delete_word_right_delete_to_end_of_line():
         text_area.load_text("01234\n56789")
         text_area.selection = Selection.cursor((0, 3))
 
-        await pilot.press("ctrl+f")
+        await pilot.press("alt+delete")
 
         assert text_area.text == "012\n56789"
         assert text_area.selection == Selection.cursor((0, 3))
@@ -465,7 +478,7 @@ async def test_delete_word_right_at_end_of_line():
         text_area.load_text("01234\n56789")
         text_area.selection = Selection.cursor((0, 5))
 
-        await pilot.press("ctrl+f")
+        await pilot.press("alt+delete")
 
         assert text_area.text == "0123456789"
         assert text_area.selection == Selection.cursor((0, 5))
@@ -477,7 +490,7 @@ async def test_delete_word_right_at_end_of_line():
         "enter",
         "backspace",
         "ctrl+u",
-        "ctrl+f",
+        "ctrl+delete",
         "ctrl+w",
         "ctrl+k",
         "ctrl+x",
@@ -563,3 +576,37 @@ async def test_paste_read_only_does_nothing():
         await pilot.pause()
 
         assert text_area.text == TEXT  # No change
+
+
+async def test_delete_to_start_of_line_deletes_newline():
+    """Test that default ctrl+u deletes newline when the cursor is at the start"""
+
+    class TextAreaApp(App):
+        def compose(self) -> ComposeResult:
+            text_area = TextArea.code_editor()
+            text_area.insert("Hello\nWorld")
+
+            yield text_area
+
+    app = TextAreaApp()
+    async with app.run_test() as pilot:
+        text_area = app.query_one(TextArea)
+
+        # Two lines to start
+        assert text_area.text == "Hello\nWorld"
+
+        # Delete the line
+        await pilot.press("ctrl+u")
+        assert text_area.text == "Hello\n"
+
+        # Delete at start, deletes the \n
+        await pilot.press("ctrl+u")
+        assert text_area.text == "Hello"
+
+        # Delete the line
+        await pilot.press("ctrl+u")
+        assert text_area.text == ""
+
+        # Delete empty is a NOOP
+        await pilot.press("ctrl+u")
+        assert text_area.text == ""